+++ crates/fake_package/src/lib.rs
@@ -1,4 +1,5 @@
 mod dependency;
+pub mod metadata;
 mod package;
 
 pub use dependency::*;