+++ crates/next_version/tests/all/pre_release.rs
@@ -3,7 +3,7 @@ use semver::Version;
 
 #[test]
 fn commit_without_semver_prefix_increments_pre_release_version() {
-    let commits = vec!["my change"];
+    let commits = ["my change"];
     let version = Version::parse("1.0.0-alpha.2").unwrap();
     let expected = Version::parse("1.0.0-alpha.3").unwrap();
     assert_eq!(version.next(commits), expected);
@@ -11,7 +11,7 @@ fn commit_without_semver_prefix_increments_pre_release_version() {
 
 #[test]
 fn commit_with_breaking_change_increments_pre_release_version() {
-    let commits = vec!["feat!: break user"];
+    let commits = ["feat!: break user"];
     let version = Version::parse("1.0.0-alpha.2").unwrap();
     let expected = Version::parse("1.0.0-alpha.3").unwrap();
     assert_eq!(version.next(commits), expected);
@@ -19,7 +19,7 @@ fn commit_with_breaking_change_increments_pre_release_version() {
 
 #[test]
 fn dot_1_is_added_to_unversioned_pre_release() {
-    let commits = vec!["feat!: break user"];
+    let commits = ["feat!: break user"];
     let version = Version::parse("1.0.0-alpha").unwrap();
     let expected = Version::parse("1.0.0-alpha.1").unwrap();
     assert_eq!(version.next(commits), expected);
@@ -27,7 +27,7 @@ fn dot_1_is_added_to_unversioned_pre_release() {
 
 #[test]
 fn dot_1_is_added_to_last_identifier_in_pre_release() {
-    let commits = vec!["feat!: break user"];
+    let commits = ["feat!: break user"];
     let version = Version::parse("1.0.0-beta.1.2").unwrap();
     let expected = Version::parse("1.0.0-beta.1.3").unwrap();
     assert_eq!(version.next(commits), expected);
@@ -35,7 +35,7 @@ fn dot_1_is_added_to_last_identifier_in_pre_release() {
 
 #[test]
 fn dot_1_is_added_to_character_identifier_in_pre_release() {
-    let commits = vec!["feat!: break user"];
+    let commits = ["feat!: break user"];
     let version = Version::parse("1.0.0-beta.1.a").unwrap();
     let expected = Version::parse("1.0.0-beta.1.a.1").unwrap();
     assert_eq!(version.next(commits), expected);